fix for threshold term

+++ newrelic-sync/src/main/java/com/ocado/pandateam/newrelic/sync/configuration/condition/terms/TermsConfiguration.java
@@ -12,7 +12,7 @@ public class TermsConfiguration {
     @NonNull
     private PriorityTerm priorityTerm;
     @NonNull
-    private Integer thresholdTerm;
+    private Float thresholdTerm;
     @NonNull
     private TimeFunctionTerm timeFunctionTerm;
